Fix garbled weather footer: %.0f fed a long from lroundf
formatWeatherLine() passed lroundf(temperature) — a long — to a %.0f
conversion, which is undefined behavior. On the ESP32-C3 (RISC-V,
riscv32-esp-elf-gcc 8.4.0) integer and floating-point varargs travel in
different registers, so the mismatch shifts every following argument.
The footer renders as garbage, e.g.

  DRIZZLE 0O RH605234...      instead of      DRIZZLE 29C RH84%

The unit character and humidity are consumed from the wrong slots, so
they are corrupted too. This is likely invisible on x86 hosts, where the
same mismatch often happens to read back something plausible.

Print the rounded value as an int instead of asking printf for a double.

+++ src/services/weather_time.cpp
@@ -198,8 +198,8 @@ void formatWeatherLine(char* out, size_t out_len) {
   if (settings::temperatureFahrenheit()) {
     temperature = temperature * 9.0f / 5.0f + 32.0f;
   }
-  snprintf(out, out_len, "%s %.0f%c RH%d%%", conditionLabel(s_weather_code),
-           lroundf(temperature), unit, s_humidity_percent);
+  snprintf(out, out_len, "%s %d%c RH%d%%", conditionLabel(s_weather_code),
+           static_cast<int>(lroundf(temperature)), unit, s_humidity_percent);
 }
 
 void formatDateTimeLine(char* out, size_t out_len) {
